half-fix a serious display bug related to unicode char printing
Due to toxic being unable to know the screen width of a given unicode character, line
printing becomes buggy when certain unicode characters are used. Rather than returning
an error and printing empty lines when a length error is detected, we now just deal
with a less-bad bug that causes the line not to wrap correctly, and the window not to
scroll the correct number of lines.

This is not ideal behaviour, but it's better than having the interface completely break,
which is particularly bad when logs are enabled, as simply closing and re-opening the
window would not fix the issue.

An ideal fix for this problem would be to calculate the correct screen-width of a line
before printing it. However this is easier said than done given toxic's hacky display