feat: config files are now tied to tox data files by default
Instead of using the same default config file for every profile
we now use the same name and path as the custom tox profile
(with a .conf file extension), and create an empty config file
if it doesn't already exist. In addition, if the user passes a
non-existent file as a run option, we now create an empty file
instead of returning an error.

Behaviour should not be changed for default profiles.